Project renaming: switch « brick » to « branch [of knowledge] » [part 1]
The idea to change the name of the project is something I've been
wanting to do for a while, as using the word "brick" to represent a kind
of knowledge base may not sound at its best.

Why a « Branch of Knowledge on Open Models » ? The idea is to place this
work in the context of the Enlightenment, in the continuity of a
gathering of knowledge led by the Encyclopédie which sought to bring
together « the infinitely varied branches of human science ».

This branch would be one of them in this field of « open models », a way
of bringing this knowledge together and making it accessible.

In addition to seeking deep historical roots, speaking of a branch rather
than a base could reinforce this link to the logic of the construction
of a science, a sum of knowledge in a field.

It could help to federate the scientific community. Open models/open
science can change the way science is built, this is also an
experimental environment to learn how science can be conducted.

Not 100% convinced on how these things should be called, nothing fully
captivates me, branch may be great and meaningful.

With the translation from french to english and the need to rename the
repository, looks like a good opportunity to do this transformation!

**Part 1:** Need to do a repository rename and change of (GitHub) links,
first step for a full rebrand to get potential opinions.